# output/openapi.yaml
openapi: 3.0.0
info:
  title: veryExpress generated api server
  description: OpenApi Spec generated by veryExpress
  version: 1.0.0
paths:
  /user:
    get:
      summary: get user
      operationId: getuser
      tags:
        - user
      parameters:
        - name: _id
          in: query
          schema:
            type: string
            format: uuid
        - name: affiliateId
          in: query
          schema:
            type: string
            format: uuid
        - name: email
          in: query
          schema:
            type: string
            format: email
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: {}
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
        '405':
          description: Method Not Allowed
        '413':
          description: Payload Too Large
        '429':
          description: Too Many Requests
        '500':
          description: Internal Server Error
        '502':
          description: Bad Gateway
        '503':
          description: Service Unavailable
    post:
      summary: post user
      operationId: postuser
      tags:
        - user
      parameters: []
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: {}
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
        '405':
          description: Method Not Allowed
        '413':
          description: Payload Too Large
        '429':
          description: Too Many Requests
        '500':
          description: Internal Server Error
        '502':
          description: Bad Gateway
        '503':
          description: Service Unavailable
  /user/{id}:
    put:
      summary: put user
      operationId: putuser
      tags:
        - user
      parameters: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: {}
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
        '405':
          description: Method Not Allowed
        '413':
          description: Payload Too Large
        '429':
          description: Too Many Requests
        '500':
          description: Internal Server Error
        '502':
          description: Bad Gateway
        '503':
          description: Service Unavailable
    patch:
      summary: patch user
      operationId: patchuser
      tags:
        - user
      parameters: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: {}
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
        '405':
          description: Method Not Allowed
        '413':
          description: Payload Too Large
        '429':
          description: Too Many Requests
        '500':
          description: Internal Server Error
        '502':
          description: Bad Gateway
        '503':
          description: Service Unavailable
    delete:
      summary: delete user
      operationId: deleteuser
      tags:
        - user
      parameters: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: {}
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
        '405':
          description: Method Not Allowed
        '413':
          description: Payload Too Large
        '429':
          description: Too Many Requests
        '500':
          description: Internal Server Error
        '502':
          description: Bad Gateway
        '503':
          description: Service Unavailable
components:
  schema: {}